services/orchestrator-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.50.0
google-genai==0.6.0
httpx==0.27.2
